Attachment-based issuer notary whitelist PoC.
To support issuer notary whitelists a state has to implement the
NotaryRestrictedState interface, which requires specifying a signing key
for whitelist.

Every transaction then has to include a signed whitelist as an
attachment.

Additionally the contract for this state has to invoke notary
whitelist-specific verification logic during verify:

RestrictedNotaryWhitelistVerifier.verify(tx)

This checks that a correct signed whitelist is attached and the transaction
notary is within the whitelist.
It also checks that during state transition the signing key remains
unmodified. However, in some cases it may be required for the contract
to specify its own logic for verifying that the signing key is not
modified, as the helper logic may not detect input-output state groups
correctly for all possible scenarios.